Clean up for PR.
