+++ open_instruct/utils.py
@@ -42,6 +42,7 @@ import subprocess
 import sys
 import threading
 import time
+from concurrent import futures
 from ctypes import CDLL, POINTER, Structure, c_char_p, c_int, c_ulong, c_void_p
 from dataclasses import dataclass, field
 from multiprocessing import resource_tracker as _rt
@@ -58,6 +59,7 @@ from dateutil import parser
 from huggingface_hub import HfApi
 from ray.util import state as ray_state
 from rich.pretty import pprint
+from tqdm import tqdm
 from transformers import MODEL_FOR_CAUSAL_LM_MAPPING, HfArgumentParser
 
 MODEL_CONFIG_CLASSES = list(MODEL_FOR_CAUSAL_LM_MAPPING.keys())
@@ -66,6 +68,29 @@ MODEL_TYPES = tuple(conf.model_type for conf in MODEL_CONFIG_CLASSES)
 logger = get_logger(__name__)
 
 DataClassType = NewType("DataClassType", Any)
+
+
+def ray_get_with_progress(ray_refs: List[ray.ObjectRef], desc: str = "Processing") -> List[Any]:
+    """Execute ray.get() with a progress bar using futures.
+
+    Args:
+        ray_refs: List of ray object references
+        desc: Description for the progress bar
+
+    Returns:
+        List of results in the same order as ray_refs
+    """
+    ray_futures = [ref.future() for ref in ray_refs]
+    results = [None] * len(ray_refs)
+
+    for future in tqdm(
+        futures.as_completed(ray_futures), total=len(ray_futures), desc=desc, bar_format="{l_bar}{bar}{r_bar}\n"
+    ):
+        idx = ray_futures.index(future)
+        results[idx] = future.result()
+
+    return results
+
 
 """
 Notes:
